+++ LTI-SPRING-REST-FLIPFIT-API-JPA/LTI-SPRING-REST-FLIPFIT-API-JPA/src/main/java/com/lti/flipfit/services/GymFlipFitAdminServiceImpl.java
@@ -96,7 +96,7 @@ public class GymFlipFitAdminServiceImpl implements GymFlipFitAdminService {
 
 	@Override
 	public GymFlipFitAdmin createAdmin(GymFlipFitAdmin admin) {
-		admin.setId(counter.incrementAndGet());
+		admin.setId((int)counter.incrementAndGet());
 		admins.add(admin);
 		adminRepository.save(admin);
 		return admin;
