+++ LTI-SPRING-REST-FLIPFIT-API-JPA/LTI-SPRING-REST-FLIPFIT-API-JPA/src/main/java/com/lti/flipfit/services/GymFlipFitNotificationServiceImpl.java
@@ -31,7 +31,7 @@ public class GymFlipFitNotificationServiceImpl implements GymFlipFitNotification
      */
     @Override
 	public GymFlipFitNotification sendNotification(GymFlipFitNotification notification) {
-		notification.setId(counter.incrementAndGet());
+		notification.setId((int)counter.incrementAndGet());
 		notifications.add(notification);
         return notification;
