+++ LTI-SPRING-REST-FLIPFIT-API-JPA/LTI-SPRING-REST-FLIPFIT-API-JPA/src/main/java/com/lti/flipfit/services/GymFlipFitOwnerServiceImpl.java
@@ -61,7 +61,7 @@ public class GymFlipFitOwnerServiceImpl implements GymFlipFitOwnerService {
      */
     @Override
     public GymFlipFitOwner saveOwner(GymFlipFitOwner owner) {
-        owner.setId(counter.incrementAndGet());
+        owner.setId((int)counter.incrementAndGet());
         owners.add(owner);
         return owner;
     }
