+++ LTI-SPRING-REST-FLIPFIT-API-JPA/LTI-SPRING-REST-FLIPFIT-API-JPA/src/main/java/com/lti/flipfit/services/GymFlipFitSchedulerServiceImpl.java
@@ -47,7 +47,7 @@ public class GymFlipFitSchedulerServiceImpl implements GymFlipFitSchedulerServic
 	 */
 	@Override
 	public GymFlipFitScheduler scheduleSlot(GymFlipFitScheduler scheduler) {
-		scheduler.setId(counter.incrementAndGet());
+		scheduler.setId((int)counter.incrementAndGet());
 		schedules.add(scheduler);
 		return scheduler;
 	}
